+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 - `Improbable.Gdk.EntityTemplate` is now mutable and exposes a set of APIs to add, remove, and replace component snapshots
     - This replaces the `Improbable.Gdk.Core.EntityBuilder` class.
     - These changes also allow you to reuse an `EntityTemplate` more than once.
+- Upgraded the project to be compatible with `2018.3.2f1`.
+- Upgraded the entities package to `0.0.12-preview.21`
+- Disabled protocol logging on Linux workers to prevent crashes. This will be reverted once the underlying issue is fixed.
 
 ### Fixed
 
@@ -27,6 +30,7 @@
 
 - Removed the `Improbable.Gdk.Core.EntityBuilder` class as it was superceded by the updated functionality in `Improbable.Gdk.Core.EntityTemplate`.
     - Removed `CreateSchemaComponentData` from each generated component as it is no longer required by the `EntityBuilder`.
+- Removed `com.unity.incrementalcompiler` package as a dependency of the `Core` package.
 
 ## `0.1.3` - 2018-11-26